# docs/swagger.yaml
definitions:
  admin_dto.Album:
    properties:
      created_at:
        type: integer
      id:
        type: string
      participant_count:
        type: integer
      title:
        type: string
      total_cards:
        type: integer
    type: object
...
  admin_dto.User:
    properties:
      created_at:
        type: integer
      description:
        type: string
      email:
        type: string
      github:
        type: string
      id:
        type: string
      is_admin:
        type: integer
      linkedin:
        type: string
      name:
        type: string
      picture:
        type: string
      updated_at:
        type: integer
      web:
        type: string
    type: object
  admin_dto.UserDetail:
    properties:
      albums:
        items:
          $ref: '#/definitions/admin_dto.Album'
        type: array
      cards:
        items:
          $ref: '#/definitions/card_model.Card'
        type: array
      user:
        $ref: '#/definitions/admin_dto.User'
    type: object
...
  card_model.Card:
    properties:
      album_id:
        type: string
      created_at:
        type: integer
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
      updated_at:
        type: integer
    type: object
...
  user_dto.UpdateUserRequest:
    properties:
      description:
        type: string
      email:
        type: string
      github:
        type: string
      linkedin:
        type: string
      name:
        type: string
      web:
        type: string
    type: object
  user_model.User:
    properties:
      created_at:
        type: integer
      description:
        type: string
      email:
        type: string
      github:
        type: string
      id:
        type: string
      is_admin:
        type: integer
      linkedin:
        type: string
      name:
        type: string
      picture:
        type: string
      update_at:
        type: integer
      web:
        type: string
    type: object
info:
  contact: {}
  description: MVP para intercambio de tarjetas coleccionables.
  title: Trading Card Album API
  version: "1.0"
paths:
...
  /api/v1/admin/users/{id}:
    get:
      consumes:
      - application/json
      description: Get user detail with albums and cards
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.UserDetail'
      security:
      - BearerAuth: []
      tags:
      - Admin
...
  /api/v1/user/{id}:
    put:
      consumes:
      - application/json
      description: Update user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_model.User'
      security:
      - BearerAuth: []
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"